# obs-filter: NVIDIA RTX AI Greenscreen filter (Background removal)