--- a/plugins/obs-filters/obs-filters-config.h.in
+++ b/plugins/obs-filters/obs-filters-config.h.in
@@ -9,3 +9,5 @@
 #endif
 
 #cmakedefine HAS_NOISEREDUCTION
+#define NVIDIA_GREENSCREEN_ENABLED @NVVFX_FOUND@
+