… plugins/obs-filters/data/locale/en-US.ini
… plugins/obs-filters/nvidia-greenscreen-filter.c
… plugins/obs-filters/nvvfx-load.h