### Description
This adds support for the Nvidia RTX Video FX sdk.
The AI Greenscreen FX with background removal is implemented.
The AI processing is quite fast and adds only 1-2 ms of rendering time for the frames. At 60 fps this means there's no lagging in the frames since they're issued every 16.7 ms.
Testers uncovered a bug in the Audio FX sdk which causes interference between the video fx and audio fx sdk, preventing the use of the video fx sdk.
This bug is fixed locally in the 3rd commit of this PR (the bugfix can be removed later when it is fixed upstream in a new sdk release).
The audio fx sdk was upgraded to v1.1.0.5 in order to enable this fix.
This fix would have been in a separate PR but the CI builds wouldn't have worked without it. So in order to enable testing, I felt better to include it here.

**Testing this PR**
People who wish to test this PR can try to retrieve artefacts (test builds) by checking the Checks tab or the Actions tab.
Direct link: https://github.com/obsproject/obs-studio/actions/runs/2711346971
(Requires a github account to be able to download the test builds.)
The Nvidia RTX video fx redist needs to be installed: https://www.nvidia.com/en-us/geforce/broadcasting/broadcast-sdk/resources/
An RTX GPU is required.

**Warning : GPU usage**
This filter impacts GPU considerably. With Unigine Heaven I observed that fps dropped from 153 fps to 134 fps (in perf mode), so about 15% drop.

**Acknowledgments:**
I'd like to thank warmly : 
- Stephan Seitz (Nvidia) who provided me a lot of guidance in the use of the Video FX sdk and without whom I wouldn't have been able to complete this project. In particular his tutoring of Nsight helped me to identify parts of the code which required improvement.
- I thank also the sdk devs at nvidia who were really helpful in debugging a tricky sdk issue. In particular Nitin Gode who solved the bug and provided a workaround.
- on obsproject side, I want to thank  @jpark37 who provided crucial help in understanding how to capture textures.
-  @WizardCM , @derrod also from obsproject helped me in testing the filter which allowed to uncover & fix several bugs.
- I want also to give credit to Xaymar who was the first to develop an obs filter for RTX Greenscreen; in particular thanks to [his code for StreamFX](https://github.com/Xaymar/obs-StreamFX/blob/master/source/nvidia/vfx/nvidia-vfx-greenscreen.cpp#L162), I understood how to pass textures from CUDA to D3D11 without doing explicitly the interop. This was a particularly tricky step.
- I should also give credit to Kounoike, whose [obs-virtualbg project ](https://github.com/kounoike/obs-virtualbg) helped me in the early stages of this project.

**Can we have smaller latency added ? Some gory details.**
I used Nsight w/ nvtx extension which has been integrated into libobs profiler by Stephan Seitz (Nvidia).
This allowed to trace visually the various calls and to optimize to the minimum the render time.
![nsys-ui_2022-04-18_17-07-25](https://user-images.githubusercontent.com/9283407/164890352-baa0d882-0323-47a4-8d69-7017ed91a7d4.png)
The graph shows that the filter .video_render calls (grey rectangles named greenscreen_render) last about 2 ms (in practice 1-2 ms); they're repeated every 16.7 ms which corresponds to 60 fps.
On the next row (cuda api), the green rectangles are the d3d11 <=> cuda transfer of textures. In between them is the AI processing.
This shows the rendering time in .video_render callback is minimized. There's no idle time. We can't go beyond 1-2 ms.


### Motivation and Context
The motivation is to have background removal done by AI without a physical green screen.
The nvidia broadcast app supports that feature (and many others) but direct integration to
obs brings some advantages:
- it allows use on more video sources, not just webcams.
- eg, Media Sources, VLC sources, Capture cards, game captures, browser sources, still pictures, Blackmagic, AJA ...
Here are some examples:

- Media Source : 
![obs64_2022-04-11_11-29-14](https://user-images.githubusercontent.com/9283407/164522059-2103e51f-22d6-4fcd-b324-597d7ec0c7ef.png)
- Picture : 
![obs64_2022-04-21_19-36-04](https://user-images.githubusercontent.com/9283407/164522084-ae5c4082-1095-4d45-a325-31d3a2cb634d.png)
- Browser Source : 
![obs64_2022-04-21_19-37-41](https://user-images.githubusercontent.com/9283407/164522088-e7ccde6f-b831-4956-8234-9553cf499f81.png)
- Webcam (Video Capture device) : 
![matt](https://user-images.githubusercontent.com/9283407/164523610-86c884e0-7284-41f0-bc5d-b84717d6dfe9.png)

### How Has This Been Tested?
Tested by myself and @WizardCM  on win 10 x64 with RTX 3070 & 3080.
Works fine.
Here are some stats:
- With Unigine Heaven I observed that fps dropped from 153 fps to 124 fps quality mode) or 134 fps (perf mode) w/ an RTX 3070, so about 15-25% drop.
- The added time to rendering is 0.5 - 1.5 ms. So at 60 fps where a frame is rendered every 16.7 ms, this is negligible.


### Types of changes
- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

### Checklist:
- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.
- [x] I have included updates to all appropriate documentation.