@pkviet 2022-04-24
> In descriptive text (not in function/variable names), always use "NVIDIA" with all capital letters.
done
> 
> In the first commit, use a verb as the first word, such as "Add".
done
> 
> Please check recommendations in this review throughout this PR and the split off PR (#6365).
will do