@pkviet 2022-06-28
**Update 06/28/22**
Improved efficiency by rendering only when there's a new frame provided by the video source.
(ex: webcam at 30 fps while obs is set to output at 60 fps)