@Fenrirthviti 2022-06-28
Tested on my system, and it's all working great for me.

One note is that we cannot provide a link to download the redistributable, per GPL, which implies that the redistributable is required for the program to be complete. The log message stating it cannot be loaded is fine, but the link to the download will need to be removed. There's also an additional space in the log, though I'm not sure if this is from the greenscreen plugin or something else, but it shows up as:

```
13:11:24.641: [NVIDIA RTX VIDEO FX]: FX disabled, redistributable not found.
13:11:24.641: Download here: https://www.nvidia.com/en-us/geforce/broadcasting/broadcast-sdk/resources/ 
13:11:24.641: 
```

Other than that, looks good to me!