@pkviet 2022-06-30
**WARNING**
AT MERGE TIME, DISCARD THE LAST TWO COMMITS OR ASK ME TO REBASE TO DROP THEM.
THEY'RE INCLUDED TO ENABLE TESTING DUE TO THIS NASTY NVIDIA AUDIO SDK BUG WHICH INTERFERES WITH VIDEO SDK.
THEY ARE ALREADY INCLUDED IN A SEPARATE RTX AUDIO PR #6372