@pkviet 2022-07-21
> I did some quick testing on a laptop with an RTX 2070. The filter functioned exactly as expected. Nothing surprising there. Adding the filter is the same as our regular process for adding a filter, so there's no added complexity from the UX standpoint. Looking at the Performance tab of Task Manager, GPU 3D usage went from about 5% to 35-45% after activating the filter at the "Quality" Mode. With the "Performance" Mode, I saw the GPU 3D usage fluctuate between 20% and 40%. When activating the filter, the GPU 3D usage briefly spikes to 100%, but then settles down to the aforementioned 35-45%.
> 
> During one session, I did get OBS into a "not responding" state when exiting OBS while the filter enabled on the laptop's built-in camera. However, since this occurred on application exit, I'm not overly concerned with this, and I wouldn't block on that observation.

on the perfs, I found out the task manager can be quite inconsistent in the figures reported. It depends on the frequency of the GPU at the time of the test. It's not like 45% of the GPU is really used in the 3d engine, as in a game.