@pkviet 2022-07-24
**Comparison with Nvidia Broadcast**
Ran some tests with Unigine Heaven.
I pitted obs + the filter against obs + Nvidia Broadcast virtual camera.
The fps are identical in performance mode; and within 3 % in quality mode.
So the two are comparable.