@pkviet 2022-07-24
> 
> Re-tested the latest build. I haven't run into the previously encountered "not responding" state, and that was the only real issue I'd encountered. Great work!

fix suggested by r1ch, glad it works.